basic unix connection working
