+++ Cargo.toml
@@ -5,8 +5,8 @@ edition = "2021"
 
 [dependencies]
 opencv = "0.94.1"
+tokio = { version = "1.28", features = ["full"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 # ffmpeg-next = "6.0"
-# tokio = { version = "1.28", features = ["full"] }
 # tokio-tungstenite = "0.19" # For WebSocket
-# serde = { version = "1.0", features = ["derive"] }
-# serde_json = "1.0"
