+++ README.md
@@ -1,0 +1,18 @@
+### Request format:
+
+```
+{
+    game: "game",
+    winner: "player A",
+    cards: [
+        {
+            player: "player A",
+            card: "card"
+        },
+        {
+            player: "player B",
+            card: "card"
+        }
+    ]
+}
+```
